# cosmo: remove calls to reflect in WCDM benchmarks