```
name                               old time/op  new time/op  delta
WCDMEN-4                           1.61ms ± 1%  1.62ms ± 1%     ~     (p=0.841 n=5+5)
WCDMENdistance-4                   10.1ms ±10%   1.6ms ± 1%  -83.73%  (p=0.008 n=5+5)
WCDME-4                             155ns ± 1%   154ns ± 0%     ~     (p=0.365 n=5+5)
WCDMEinv-4                          176ns ± 1%   176ns ± 1%     ~     (p=1.000 n=5+5)
WCDMComovingDistance-4              357µs ± 1%   355µs ± 1%     ~     (p=0.151 n=5+5)
WCDMComovingTransverseDistance-4    356µs ± 1%   357µs ± 1%     ~     (p=0.548 n=5+5)
WCDMLuminosityDistance-4            356µs ± 1%   356µs ± 1%     ~     (p=0.841 n=5+5)
WCDMLookbackTime-4                  375µs ± 2%   363µs ± 0%   -3.35%  (p=0.008 n=5+5)
WCDMNComovingDistance-4             3.59s ± 1%   3.53s ± 0%   -1.69%  (p=0.008 n=5+5)
WCDMNLuminosityDistance-4           3.64s ± 0%   3.54s ± 0%   -2.69%  (p=0.008 n=5+5)
WCDMNE-4                           10.4ms ± 8%   1.6ms ± 1%  -84.17%  (p=0.008 n=5+5)
WCDMComovingDistanceOM-4            991ns ± 5%   155ns ± 0%  -84.36%  (p=0.008 n=5+5)
WCDMLookbackTimeOM-4               1.32µs ± 3%  0.49µs ± 1%  -62.85%  (p=0.008 n=5+5)
WCDMComovingDistanceFlat-4         1.52µs ± 5%  0.63µs ± 1%  -58.34%  (p=0.008 n=5+5)
WCDMComovingDistancePositiveOk0-4   208µs ± 1%   212µs ± 1%   +1.55%  (p=0.016 n=5+5)
```